requirement.txt
fastapi
uvicorn
sqlalchemy
alembic
passlib
bcrypt
python-jose
python-multipart